fixed 6

+++ 0x00-hello_world/6-size.c
@@ -9,7 +9,7 @@ int main(void)
 {
 	printf("Size of a char: %lu byte(s)", sizeof(char));
 	printf("\nSize of an int: %lu byte(s)", sizeof(int));
-	printf("\nSize of a long int: %lu byte(s)", sizeof(int));
+	printf("\nSize of a long int: %lu byte(s)", sizeof(short));
 	printf("\nSize of a long long int: %lu byte(s)", sizeof(long));
 	printf("\nSize of a float: %lu byte(s)", sizeof(float));
 	return (0);
